+++ Editor/InternalBridge/AssemblyInfo.cs
@@ -1,5 +1,6 @@
 using System.Runtime.CompilerServices;
 
+[assembly: InternalsVisibleTo("Unity.2D.Common.Tests.EditorTests")]
 [assembly: InternalsVisibleTo("Unity.2D.Animation.Editor")]
 [assembly: InternalsVisibleTo("Unity.2D.Animation.Tests.EditorTests")]
 [assembly: InternalsVisibleTo("Unity.2D.SpriteShape.Editor")]
@@ -8,3 +9,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.2D.Enhancers.Editor")]
 [assembly: InternalsVisibleTo("Unity.2D.Enhancers.Tests.Editor.Tests")]
 [assembly: InternalsVisibleTo("Unity.2D.Muse.Editor")]
+[assembly: InternalsVisibleTo("Unity.2D.Tooling.Editor")]