+++ Path/Editor/AssemblyInfo.cs
@@ -2,3 +2,4 @@ using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Unity.2D.Common.Path.Tests.EditorTests")]
 [assembly: InternalsVisibleTo("Unity.2D.SpriteShape.Editor")]
+[assembly: InternalsVisibleTo("Unity.RenderPipelines.Universal.Editor")]