+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [12.0.0] - 2025-07-17
+### Changed
+- Update minimum Unity version.
+- Improved outline generation.
+
+
 ## [11.0.1] - 2025-05-16
 ### Changed
 - Added support code for PolySpatial.